created RL_posM folder
created a folder contains a new environment for RL. now we use (pos, M) as the state.

Note TW: updating the M_biased by biasing original M with total_bias could be non-trivial and the M_biased gets unstable when we discard the real and negative part. even if we normalized the M_biased. maybe we should use state (pos, K) instead?