Added posM folder. Runnable. Not trained yet.
We now propagate the system using Markov matrix M. The re-defined RL env is as follow: state is a tuple made of (pos, M) with shape ([1,], [N,N]], and the action is a vector range(0, N-1) with shape [N], representing the gaussian position on the axis. Note that this is similar to one-hot but we can have multiple gaussians at the same position.